# gbvmis/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a gbvmis api",
        "title": "SGBV Police API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080/api",
    "basePath": "/",
    "paths": {
...
        "/police-post": {
            "post": {
                "description": "Creates a new police post entry in the system and returns the created record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Police Posts"
                ],
                "summary": "Create a new police post record",
                "parameters": [
                    {
                        "description": "PolicePost data to create",
                        "name": "Policepost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PolicePost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created police post record",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad request due to invalid input",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when creating police post",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/police-post/{id}": {
            "get": {
                "description": "Fetches a policePost record based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Police Posts"
                ],
                "summary": "Retrieve a single policePost record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PolicePost ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PolicePost retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "PolicePost not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when retrieving PolicePost",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the details of a policePost record based on the provided ID and request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Police Posts"
                ],
                "summary": "Update an existing policePost record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PolicePost ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PolicePost data to update",
                        "name": "policePost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdatePolicePostPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PolicePost updated successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Invalid input or empty request body",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "PolicePost not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when updating policePost",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a PolicePost record based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Police Posts"
                ],
                "summary": "Delete a PolicePost record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PolicePost ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PolicePost deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "PolicePost not found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Server error when deleting policePost",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/police-posts": {
            "get": {
                "description": "Fetches all policePost records with pagination support.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Police Posts"
                ],
                "summary": "Retrieve a paginated list of policePosts",
                "responses": {
                    "200": {
                        "description": "policePosts retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve policePosts",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/police-posts/search": {
            "get": {
                "description": "Retrieves a paginated list of policePosts based on search criteria.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Police Posts"
                ],
                "summary": "Search for policePosts with pagination",
                "responses": {
                    "200": {
                        "description": "PolicePosts retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve PolicePosts",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.UpdatePolicePostPayload": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Case": {
            "type": "object",
            "properties": {
                "case_number": {
                    "type": "string"
                },
                "charges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Charge"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "date_opened": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "officer": {
                    "$ref": "#/definitions/models.PoliceOfficer"
                },
                "officer_id": {
                    "type": "integer"
                },
                "policePost": {
                    "$ref": "#/definitions/models.PolicePost"
                },
                "police_post_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "suspect_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "victims": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Victim"
                    }
                }
            }
        },
        "models.Charge": {
            "type": "object",
            "properties": {
                "case_id": {
                    "type": "integer"
                },
                "charge_title": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "severity": {
                    "description": "e.g., Felony, Misdemeanor",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.PoliceOfficer": {
            "type": "object",
            "properties": {
                "badge_no": {
                    "type": "string"
                },
                "cases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Case"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "post_id": {
                    "type": "integer"
                },
                "rank": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Role"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.PolicePost": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "officers": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PoliceOfficer"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Role": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Victim": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cases": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Case"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "dob": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "nin": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        }
    }
}